refactoring: 구간 삭제 인수 테스트 리펙토링
