+++ src/test/java/nextstep/subway/section/SectionAcceptanceTest.java
@@ -1,21 +1,17 @@
 package nextstep.subway.section;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.line.dto.LineRequest;
-import nextstep.subway.line.dto.LineResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
 
 import static nextstep.subway.line.LineRequestStep.*;
-import static nextstep.subway.station.StationRequestStep.지하철역_등록되어_있음;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
-import static org.springframework.http.HttpStatus.CREATED;
+import static nextstep.subway.section.SectionRequestStep.*;
+import static nextstep.subway.section.SectionVerificationStep.*;
+import static nextstep.subway.station.StationRequestStep.*;
 
 @DisplayName("지하철 구간 관련 테스트")
 class SectionAcceptanceTest extends AcceptanceTest {
@@ -64,22 +60,8 @@ class SectionAcceptanceTest extends AcceptanceTest {
     void delete() {
         지하철_구간_등록되어_있음(신분당선, LineRequest.of(강남역, 역삼역, 10));
 
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .when().delete("/lines/{lineId}/sections/", 신분당선)
-                .then().log().all().extract();
+        ExtractableResponse<Response> response = 지하철_구간_삭제요청(신분당선);
 
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
-
-
-    private void 지하철_구간_등록실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(BAD_REQUEST.value());
-    }
-
-    private void 지하철_구간_등록됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(CREATED.value());
-        assertThat(response.body().as(LineResponse.class).getSections().size()).isEqualTo(1);
+        지하철_구간_삭제됨(response);
     }
 }
